[DDW-197] Moves ThemeProvider into components/HOC
